Port to Bash on Windows
Changing the shell didn't work with Bash on Windows. Instead, assume
sh.exe is in the path and invoke it from a batch script.

Could be improved by putting sh.exe in your path or doing more checking.
Better solution would be to port shell code to python and invoke
directly from vim.