Re-use the thesaurus window if it still exists
Store the buffer number for the thesaurus window and jump back to it if
it still exists on next invocation.

If it's open in another tab, bufwinnr will return -1 (because it's not
visible). That means you'll get it open in two tabs, but that sounds
like that you'd want (rather than jumping to the other tab).